refactor: move Feedback model to Models directory
